@@ -1,18 +1,21 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.css';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import store from "./redux/store";
 import {Provider} from "react-redux";
 
 
-ReactDOM.render(
-    //<React.StrictMode>
-    <Provider store={store}>
-        <App />
-    </Provider>,
-    //</React.StrictMode>,
-    document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+  <Provider store={store}>
+    <App />
+  </Provider>
+  </React.StrictMode>
 );
 
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
