@@ -4,13 +4,13 @@ import TypingPanel from "./typingPanel/TypingPanel";
 import ConfirmPanel from "./ConfirmPanel/ConfirmPanel";
 
 
-const Instructions = () => {
+const Instructions = ({ winningIndex }) => {
 
     return (
         <div className="col-lg-8 col-md-8 instructions-body">
             <div className="row">
                 <TypingPanel />
-                <ConfirmPanel />
+                <ConfirmPanel winningIndex={winningIndex}/>
             </div>
         </div>
     )
